requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiomysql
python-dotenv
pydantic

